refactor: 카테고리 예시 데이테 분리
